Add the gem for user authentication with Google (#4)
* Add the gem for user authentication with Google

Users should log in with their Google account. Adds the OmniAuth Google
OAuth2 gem in order to authenticate with Google.

Implement: #1

* Extract user auth validation to the controller

The command which returns authentication information also verifies if
the auth data are valid with a specific user. Extracts the behavior to
the controller to permit to have more auth types that use the auth user
command.

Implement: #1

* Implement Google Omniauth to authenticate users

Users need to authenticate through google oauth2, implements the
configuration to permit this auth.

Implement: #1

* Convert user as json during auth

* remove foreign key

* Render HTML instead of json with Omniauth session

Implement: #1

Co-authored-by: mickmister <mjkochell@gmail.com>